+++ filenames.auto.gni
@@ -273,6 +273,7 @@ auto_filenames = {
     "lib/common/type-utils.ts",
     "lib/common/web-view-methods.ts",
     "lib/common/webpack-globals-provider.ts",
+    "lib/common/webpack-provider.ts",
     "lib/renderer/api/context-bridge.ts",
     "lib/renderer/api/crash-reporter.ts",
     "lib/renderer/api/desktop-capturer.ts",
@@ -294,7 +295,6 @@ auto_filenames = {
     "lib/renderer/web-view/web-view-element.ts",
     "lib/renderer/web-view/web-view-impl.ts",
     "lib/renderer/web-view/web-view-init.ts",
-    "lib/renderer/webpack-provider.ts",
     "lib/renderer/window-setup.ts",
     "package.json",
     "tsconfig.electron.json",
@@ -315,6 +315,7 @@ auto_filenames = {
     "lib/common/reset-search-paths.ts",
     "lib/common/type-utils.ts",
     "lib/common/webpack-globals-provider.ts",
+    "lib/common/webpack-provider.ts",
     "lib/renderer/api/context-bridge.ts",
     "lib/renderer/api/crash-reporter.ts",
     "lib/renderer/api/desktop-capturer.ts",
@@ -326,8 +327,18 @@ auto_filenames = {
     "lib/renderer/ipc-renderer-internal-utils.ts",
     "lib/renderer/ipc-renderer-internal.ts",
     "lib/renderer/remote/callbacks-registry.ts",
-    "lib/renderer/webpack-provider.ts",
     "lib/worker/init.ts",
+    "package.json",
+    "tsconfig.electron.json",
+    "tsconfig.json",
+    "typings/internal-ambient.d.ts",
+    "typings/internal-electron.d.ts",
+  ]
+
+  asar_bundle_deps = [
+    "lib/asar/fs-wrapper.ts",
+    "lib/asar/init.ts",
+    "lib/common/webpack-provider.ts",
     "package.json",
     "tsconfig.electron.json",
     "tsconfig.json",
