+++ lib/asar/init.ts
@@ -1,0 +1,3 @@
+import { wrapFsWithAsar } from './fs-wrapper';
+
+wrapFsWithAsar(require('fs'));
