--- lib/common/asar_init.js
@@ -1,4 +1,0 @@
-'use strict';
-
-// Monkey-patch the fs module.
-require('electron/js2c/asar').wrapFsWithAsar(require('fs'));
